requirements.txt
django==3.2.20
djangorestframework==3.14.0
djongo==1.3.6
pymongo==3.12.3
Pillow==11.2.1
django-cors-headers==4.3.1